refactor(cli): extract compare.py run resolution to _compare_runs
Moves _detect_by + _resolve_runs out of
amx/cli_support/commands/compare.py into a focused
amx/cli_support/_compare_runs.py, with the two supporting constants
_BY_DIMENSIONS and _BY_TO_RUN_KEY.

_resolve_runs reads history_store off the compare module via lazy
import inside the function, so existing tests that
patch('amx.cli_support.commands.compare.history_store', ...) keep
affecting the call site unchanged.

compare.py re-exports all four names. compare.py: 2264 -> 2178 LOC (-86).
